Move qlog to internal/qlog and use qtest.Logger(t)
This moves the qlog package to internal/qlog and replaces the old
qlog.Logger(t) method with the qtest.Logger(t) which had the same
functionality.
