+++ cmd/scm/main.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 
 	"github.com/quickfeed/quickfeed/database"
+	"github.com/quickfeed/quickfeed/internal/qlog"
 	"github.com/quickfeed/quickfeed/qf"
-	"github.com/quickfeed/quickfeed/qlog"
 	"github.com/quickfeed/quickfeed/scm"
 
 	"github.com/urfave/cli"
