+++ internal/qtest/logger_helper.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
 )
 
 func Logger(t *testing.T) *zap.SugaredLogger {
@@ -12,9 +13,13 @@ func Logger(t *testing.T) *zap.SugaredLogger {
 	if os.Getenv("LOG") == "" {
 		return zap.NewNop().Sugar()
 	}
-	logger, err := zap.NewDevelopment()
+	cfg := zap.NewDevelopmentConfig()
+	// add colorization
+	cfg.EncoderConfig.EncodeLevel = zapcore.CapitalColorLevelEncoder
+	// we only want stack trace enabled for panic level and above
+	logger, err := cfg.Build(zap.AddStacktrace(zapcore.PanicLevel))
 	if err != nil {
-		t.Fatal(err)
+		t.Fatalf("cannot initialize logger: %v", err)
 	}
 	return logger.Sugar()
 }
